move SyncToLocalNb consumer and producer structs into local_nb module
