+++ src/local_nb.rs
@@ -8,6 +8,7 @@ use crate::sync::{BulkPipeError, PipeError};
 
 pub mod consumer;
 pub mod producer;
+pub mod sync_to_local_nb;
 
 /// A `Consumer` consumes a potentially infinite sequence, one item at a time.
 ///
